DTO changed of offset key because for cities having timezone as float for example delhi 5.30, therre was contract breach as third party API was giving offset as float

+++ app/src/main/java/com/daniloprado/weather/data/dto/ForecastDto.java
@@ -10,7 +10,7 @@ public class ForecastDto implements Serializable {
     public double latitude;
     public double longitude;
     public String timezone;
-    public Integer offset;
+    public Double offset;
     public Currently currently;
     public Daily daily;
